services/kafka_timescale_sink/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.5
pydantic==2.7.1